pyproject.toml
[tool.poetry]
name = "speech-analysis-tool"
version = "0.1.0"
description = ""
authors = ["Konrad Z <konradz321@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "4.10.0.84"
ffmpeg-python = "0.2.0"
SpeechRecognition = "~3.9"
easyocr = "1.7.2"
black = "^24.8.0"
requests = "2.32.3"
openai = "1.50.2"
pyhamcrest = "^2.1.0"
pytest = "^8.3.3"
stanza = "^1.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
